03-ai-agent-for-devops/code/07/requirements.txt
python-dotenv>=1.2.1
langchain>=1.0.5
langchain-google-genai>=3.0.3
